move mocked_datscha_web into dedicated pytest plugin
