+++ tests/datscha_web/conftest.py
@@ -7,7 +7,6 @@ from pytest import MonkeyPatch
 
 from mex.common.ldap.connector import LDAPConnector
 from mex.common.ldap.models.person import LDAPPerson
-from mex.datscha_web.connector import DatschaWebConnector
 from mex.datscha_web.models.item import DatschaWebItem
 from mex.datscha_web.settings import DatschaWebSettings
 
@@ -106,28 +105,6 @@ def datscha_web_item_without_contributors(
             "Zentrale Stelle für die Verarbeitung": None,
         }
     )
-
-
-@pytest.fixture
-def mocked_datscha_web(
-    monkeypatch: MonkeyPatch,
-    datscha_web_item: DatschaWebItem,
-    datscha_web_item_without_contributors: DatschaWebItem,
-) -> None:
-    """Mock the datscha web connector to return dummy data."""
-    mock_items = {
-        "fake://17": datscha_web_item,
-        "fake://92": datscha_web_item_without_contributors,
-    }
-    monkeypatch.setattr(
-        DatschaWebConnector,
-        "__init__",
-        lambda self: setattr(self, "session", MagicMock()),
-    )
-    monkeypatch.setattr(
-        DatschaWebConnector, "get_item_urls", lambda _: iter(mock_items)
-    )
-    monkeypatch.setattr(DatschaWebConnector, "get_item", lambda _, n: mock_items[n])
 
 
 @pytest.fixture
